fix(SplitterLayout - TypeScript): extend `children` type to allow conditional rendering
