+++ packages/main/src/components/SplitterLayout/useConcatSplitterElements.tsx
@@ -2,9 +2,10 @@ import type { CSSProperties, ReactElement } from 'react';
 import { Children, cloneElement, isValidElement, useMemo } from 'react';
 import { Splitter } from '../Splitter/index.js';
 import type { SplitterElementPropTypes } from '../SplitterElement/index.js';
+import type { SplitterLayoutPropTypes } from './types.js';
 
 interface ConcatSplitterElements {
-  children: ReactElement<SplitterElementPropTypes> | ReactElement<SplitterElementPropTypes>[];
+  children: SplitterLayoutPropTypes['children'];
   width: CSSProperties['width'];
   height: CSSProperties['height'];
   vertical: boolean;
